changed the bash script so you specify the first argument as the library you want to call Main in. i.e. application or editor.

+++ run.sh
@@ -1,3 +1,7 @@
+if [ $# -eq 0 ]; then
+    echo "No arguments provided. Please specify the library on which Main is run. i.e application or editor"
+    exit 1
+fi
 mkdir -p bin
 cd ./bin/
 rm -rf *
